Order languages alphabetically on /code
This puts the language sections in order as suggested in:

https://github.com/graphql/graphql.github.io/pull/332

Additionally, it adds small tables of contents at the top of each major
section to make it easier to locate and jump to specific language
implementations. I maintained a special call-out to the JavaScript
reference implementations because we want them to be fairly prominent
(this is why they were at the top in originally).

Manually tested that the links work by running a local build. Note that
there is some fragility here in that the links are hand-constructed and
depend on the ordering within the page (ie. the server JavaScript link
it "#javascript" but the client one is "#javascript-1" because it
appears further down the page); wanted to do the simplest thing for now
and deal with any breakage later IFF it ever happens.