Update Tutorial-Authentication.md
## Background 
The current GraphQL Authentication and Express Middleware documentation states the "request" object is passed as the second argument in a resolver.

However, the express-graphql (v0.7.1 ) documentation states the "request" object is passed as the third argument to resolvers. See here: https://github.com/graphql/express-graphql#combining-with-other-express-middleware

I've tested this and it appears the express-graphql documentation is correct.

## Proposed File Change
I've updated this page to match the express-graphql documentation.

Please let me know if you have any questions.