New url for swapi-graphql redirects
There's will be a new build for swapi-graphql at netlify once https://github.com/graphql/swapi-graphql/pull/162 is merged, we can merge this to take it live 🌍 

I'm pretty sure the order of the last two rules will be fine. If the last one doesn't work, it will still work

```
/swapi-graphql/graphql https://swapi-graphql.netlify.com/.netlify/functions/index 200
```
this just 200 redirects (non permanent?) which essentially is netlify terms for what nginx is a proxy_path

so this will just make `graphiql.org/swapi-graphql/graphql` an alias for the lambda endpoint. see the PR's develop preview before it's merged if you want to see that default lambda path in action for yourself