Move CoC to footer, add to community page
The header menu had too many items and was causing strange wrapping with the Algolia search box.
This moves the link to the CoC to the footer in a logical place, and also adds a link to the
Community Resources page.

It also adds a link to GitHub to the community page.

Signed-off-by: Brian Warner <brian@bdwarner.com>